+++ changelog.md
@@ -2,6 +2,7 @@
 **<span style="color:#56adda">0.0.9-beta1</span>**
 - Add support for the h264_vaapi encoder
 - Add support for the h264_nvenc encoder
+- Add support for the hevc_nvenc encoder
 
 **<span style="color:#56adda">0.0.8</span>**
 - update ffmpeg helper library to latest version